add separate composite module
